refactor(cosim): lift SPI-flash GPU structs to shared module (Stage C prep)
Move FlashState / FlashDinParams / FlashModelParams from `metal.rs` into the
agnostic `mod.rs` shared GPU-struct region (gated to GPU builds), with
`size_of` ABI guards (48 / 24 / 32 bytes), mirroring the Stage B IO-struct
lift. Unblocks the CUDA/HIP flash-kernel port + the CpuBackend flash oracle
without duplicating the ABI. Metal bit-identical (behaviour-neutral relocation);
no-feature + Metal builds green.

Co-developed-by: Claude Code v2.1.183 (claude-opus-4-8)
